+++ yzy_terminal_agent/database/api.py
@@ -1,5 +1,9 @@
+import logging
+from common import constants
 from yzy_terminal_agent.database.models import *
 from yzy_terminal_agent.database import BaseTableCtrl
+
+logger = logging.getLogger(__name__)
 
 
 class YzyVoiTerminalTableCtrl(BaseTableCtrl):
@@ -13,16 +17,25 @@ class YzyVoiTerminalTableCtrl(BaseTableCtrl):
         self.db.session.commit()
 
     def delete_terminal_by_mac(self, mac):
-        qry = self.model_query(YzyVoiTerminal).filter_by(mac=mac)
-        terminal_uuid = qry.first().uuid
-        qry.delete()
-        self.db.session.commit()
-        return terminal_uuid
+        qry = self.model_query(YzyVoiTerminal).filter_by(mac=mac, deleted=False).first()
+        if qry:
+            terminal_uuid = qry.uuid
+            qry.delete()
+            self.db.session.commit()
+            return terminal_uuid
 
     def update_terminal_by_mac(self, **kwargs):
         kwargs['updated_at'] = datetime.datetime.utcnow()
-        self.model_query(YzyVoiTerminal).filter_by(mac=kwargs['mac']).update(kwargs)
+        self.model_query(YzyVoiTerminal).filter_by(mac=kwargs['mac'], deleted=False).update(kwargs)
         self.db.session.commit()
+
+    def update_terminal_status_by_mac(self, **kwargs):
+        terminal = self.model_query(YzyVoiTerminal).filter_by(mac=kwargs['mac']).first()
+        status = kwargs.get("status", 0)
+        if terminal and terminal.status != status:
+            kwargs['updated_at'] = datetime.datetime.utcnow()
+            terminal.update(kwargs)
+            self.db.session.commit()
 
     def reset_group_uuid(self, group_uuid):
         update_data = {
@@ -33,11 +46,11 @@ class YzyVoiTerminalTableCtrl(BaseTableCtrl):
         self.db.session.commit()
 
     def select_terminal_by_mac(self, mac):
-        qry = self.model_query(YzyVoiTerminal).filter_by(mac=mac)
+        qry = self.model_query(YzyVoiTerminal).filter_by(mac=mac, deleted=False)
         return qry.first()
 
     def select_terminal_by_group_uuid(self, group_uuid):
-        return self.model_query(YzyVoiTerminal).filter_by(group_uuid=group_uuid)
+        return self.model_query(YzyVoiTerminal).filter_by(group_uuid=group_uuid, deleted=False)
 
     def select_all_terminal(self):
         return self.model_query(YzyVoiTerminal).filter_by()
@@ -48,6 +61,11 @@ class YzyVoiTerminalTableCtrl(BaseTableCtrl):
 
     def add_torrent_task(self, values):
         pass
+
+    def update_terminal_status(self, terminal, status):
+        terminal.status = status
+        self.db.session.add(terminal)
+        self.db.session.commit()
 
 
 class YzyAdminUserTableCtrl(BaseTableCtrl):
@@ -65,7 +83,7 @@ class YzyNetworkIpCtrl(BaseTableCtrl):
 
     def select_controller_image_ip(self):
         # 1. get controller uuid
-        qry_node = self.model_query(YzyNode).filter_by(type=1).first()
+        qry_node = self.model_query(YzyNode).filter(YzyNode.type.in_((1, 3))).first()
         if qry_node and qry_node.uuid:
             controller_uuid = qry_node.uuid
         else:
@@ -96,12 +114,33 @@ class YzyVoiTorrentTaskTableCtrl(BaseTableCtrl):
         self.db.session.commit()
 
     def select_task_by_torrent_id(self, torrent_id):
-        qry = self.model_query(YzyVoiTorrentTask).filter_by(torrent_id=torrent_id)
+        qry = self.model_query(YzyVoiTorrentTask).filter_by(torrent_id=torrent_id, deleted=False)
         return qry.first()
 
     def select_all_task(self):
         return self.model_query(YzyVoiTorrentTask).filter(YzyVoiTorrentTask.status.in_([0,1]),
                                                           YzyVoiTorrentTask.deleted==False).all()
+
+    def select_upload_task_all(self):
+        return self.model_query(YzyVoiTorrentTask).filter(YzyVoiTorrentTask.type==constants.BT_UPLOAD_TASK,
+                                                            YzyVoiTorrentTask.deleted==False).all()
+
+    def select_terminal_bt_task(self, terminal_mac, torrent_name):
+        return self.model_query(YzyVoiTorrentTask).filter_by(terminal_mac=terminal_mac,torrent_name=torrent_name,
+                                                            deleted=False).all()
+
+    def select_terminal_bt_upload_task(self, terminal_mac, torrent_name):
+        return self.model_query(YzyVoiTorrentTask).filter_by(
+                terminal_mac=terminal_mac,torrent_name=torrent_name,type=constants.BT_UPLOAD_TASK)\
+                .order_by(YzyVoiTorrentTask.id.desc()).first()
+
+    def select_terminal_all_task(self, terminal_mac):
+        return self.model_query(YzyVoiTorrentTask).filter_by(terminal_mac=terminal_mac, deleted=False).all()
+
+    def delete_tasks(self, tasks):
+        for task in tasks:
+            task.delete()
+            self.db.session.commit()
 
     def update_task_values(self, task, values):
         task.update(values)
@@ -115,11 +154,20 @@ class YzyVoiTerminalToDesktopsCtrl(BaseTableCtrl):
 
     def get_terminal_to_desktop(self, terminal_uuid, desktop_uuid):
         qry = self.model_query(YzyVoiTerminalToDesktops).filter_by(terminal_uuid=terminal_uuid,
-                                                                   desktop_group_uuid=desktop_uuid).first()
+                                                                   desktop_group_uuid=desktop_uuid, deleted=False).first()
+        return qry
+
+    def get_desktop_by_terminal_mac(self, terminal_mac):
+        qry = self.model_query(YzyVoiTerminalToDesktops).filter_by(terminal_mac=terminal_mac).all()
         return qry
 
     def delete_all_bind_by_terminal(self, terminal_uuid):
-        qry = self.model_query(YzyVoiTerminalToDesktops).filter_by(terminal_uuid=terminal_uuid)
+        qry = self.model_query(YzyVoiTerminalToDesktops).filter_by(terminal_uuid=terminal_uuid, deleted=False)
+        qry.delete()
+        self.db.session.commit()
+
+    def delete_all_bind_by_terminal_mac(self, terminal_mac):
+        qry = self.model_query(YzyVoiTerminalToDesktops).filter_by(terminal_mac=terminal_mac, deleted=False)
         qry.delete()
         self.db.session.commit()
 
@@ -128,3 +176,70 @@ class YzyVoiTerminalToDesktopsCtrl(BaseTableCtrl):
         self.db.session.add(task)
         self.db.session.commit()
 
+
+class YzyVoiDeviceInfoTableCtrl(BaseTableCtrl):
+    def __init__(self, db):
+        super(YzyVoiDeviceInfoTableCtrl, self).__init__(db)
+
+    def get_device_by_uuid(self, disk_uuid):
+        qry = self.model_query(YzyVoiDeviceInfo).filter_by(uuid=disk_uuid,deleted=False).first()
+        return qry
+
+    def delete_all_bind_by_terminal(self, terminal_uuid):
+        qry = self.model_query(YzyVoiTerminalToDesktops).filter_by(terminal_uuid=terminal_uuid, deleted=False)
+        qry.delete()
+        self.db.session.commit()
+
+    def update_task_values(self, task, values):
+        task.update(values)
+        self.db.session.add(task)
+        self.db.session.commit()
+
+
+class YzyVoiDesktopGroupTableCtrl(BaseTableCtrl):
+    def __init__(self, db):
+        super(YzyVoiDesktopGroupTableCtrl, self).__init__(db)
+
+    def get_desktop_group_by_template(self, template_uuid):
+        qry = self.model_query(YzyVoiDesktopGroup).filter_by(template_uuid=template_uuid, deleted=False).first()
+        return qry
+
+
+class YzyHaInfoTableCtrl(BaseTableCtrl):
+    def __init__(self, db):
+        super(YzyHaInfoTableCtrl, self).__init__(db)
+
+    def get_ha_info_first(self):
+        qry = self.model_query(YzyHaInfo).filter_by(deleted=False).first()
+        return qry
+
+
+class YzyTerminalPerformanceCtrl(BaseTableCtrl):
+    def __init__(self, db):
+        super(YzyTerminalPerformanceCtrl, self).__init__(db)
+
+    def select_performance_to_uuid(self, terminal_uuid):
+        qry = self.model_query(YzyVoiTerminalPerformance).filter_by(deleted=False, terminal_uuid=terminal_uuid).first()
+        return qry
+
+    def add_performance(self, values):
+        performance = YzyVoiTerminalPerformance()
+        performance.update(values)
+        self.db.session.add(performance)
+        self.db.session.commit()
+
+    def update_performance_info(self, task, values):
+        task.update(values)
+        self.db.session.add(task)
+        self.db.session.commit()
+
+
+class YzyTerminalHardwareCtrl(BaseTableCtrl):
+    def __init__(self, db):
+        super(YzyTerminalHardwareCtrl, self).__init__(db)
+
+    def add_hard_ware(self, values):
+        hard_ware = YzyVoiTerminalHardWare
+        hard_ware.update(values)
+        self.db.session.add(hard_ware)
+        self.db.session.commit()
